Moved ChangeCard to views and added @ imports
