+++ src/views/ChangeCard.tsx
@@ -3,11 +3,11 @@ import * as TruffleContract from "truffle-contract";
 import * as Web3 from "web3";
 
 const ChoreographyContract = TruffleContract(require("../../build/contracts/Choreography.json"));
-import IChoreography, { States } from "../contract-interfaces/IChoreography";
-import User from "../util/User";
-import MessageHistory, { IMessageHistoryEntry } from "./MessageHistory";
-import StackedDate from "./StackedDate";
-import StackedUser from "./StackedUser";
+import MessageHistory, { IMessageHistoryEntry } from "@/components/MessageHistory";
+import StackedDate from "@/components/StackedDate";
+import StackedUser from "@/components/StackedUser";
+import IChoreography, { States } from "@/contract-interfaces/IChoreography";
+import User from "@/util/User";
 
 const changeCardStyles = require("./ChangeCard.css");
 
@@ -41,13 +41,22 @@ export default class ChangeCard extends React.Component<IChangeCardProps, IChang
   }
 
   public async componentWillMount() {
-    const instance: IChoreography = await this.initializeContract();
+    let timestamp: Date = new Date();
+    let publicKey: string = "";
+    let diff: string = "";
+    let state: States = States.READY;
 
-    // Get current change values
-    const timestamp: Date = new Date(await instance.timestamp() * 1000);
-    const publicKey = await instance.proposer();
-    const diff: string = await instance.diff();
-    const state: States = await instance.state();
+    try {
+      const instance: IChoreography = await this.initializeContract();
+
+      // Get current change values
+      timestamp = new Date(await instance.timestamp() * 1000);
+      publicKey = await instance.proposer();
+      diff = await instance.diff();
+      state = await instance.state();
+    } catch (e) {
+      // do nothing
+    }
 
     const proposer = await User.build(publicKey, "friedow");
     const user2 = await User.build("x", "MaximilianV");
@@ -96,6 +105,7 @@ export default class ChangeCard extends React.Component<IChangeCardProps, IChang
       from: this.props.web3.eth.accounts[0],
       gas: 1000000,
     });
+    ChoreographyContract.new(["friedow", "christian@friedow.com"]);
 
     // Initialize contract instance
     let instance: IChoreography;
