+++ tsconfig.json
@@ -14,7 +14,13 @@
     "alwaysStrict": true,
     "noEmitOnError": true,
     "noUnusedLocals": true,
-    "noUnusedParameters": false
+    "noUnusedParameters": false,
+    "baseUrl": ".",
+    "paths": {
+      "@/*": [
+        "src/*"
+      ]
+    }
   },
   "include": [
     "src",
