+++ tslint.json
@@ -16,6 +16,8 @@
       false
     ],
     "no-var-requires": false,
-    "object-literal-sort-keys": false
+    "object-literal-sort-keys": false,
+    "no-submodule-imports": false,
+    "no-implicit-dependencies": false
   }
 }
