+++ webpack.config.ts
@@ -124,6 +124,9 @@ const buildConfig: webpack.Configuration = {
   ],
   resolve: {
     extensions: [".ts", ".tsx", ".js", ".jsx"],
+    alias: {
+      "@": path.resolve(__dirname, "src"),
+    },
   },
 };
